==> scripts/7b_Predictors_selection.R <==
#------------- Description ---------------------
# Purpose: 
# This script aims to compute select predictors. 

# Author: Marieke Schultz

# Date script created: 16/11/2025

#------------- Setting up ------------------
# Remove existing objects
rm(list = ls())

# Set current working directory
setwd("/media/marieke/Shared/Chap-1/Model/Scripts/Chap_1_2018-2024")

# Libraries
library(corrplot)
library(dplyr)
library(exactextractr)
library(sf)
library(raster)
library(ncdf4)
library(lubridate)
library(terra)
library(stringr)
library(pMEM)



# Load functions
source("./utils/Fct_Data-Prep.R")

#------------- Load data ------------------
# predictors_tr_v.1.3 ----
pred <- st_read("./data/processed_data/predictors/predictors_tr_v1.3.gpkg")
names(pred)
# mtdt_7_sel_v1.1 ----
mtdt <- st_read("./data/processed_data/Mtdt/mtdt_7_sel_v1.1.gpkg") # no lockdown, automn-winter, > 50m sites

#---------------- REMOVE ROWS BASED ON SITE SELECTION -----

# Remove pred rows not in mtdt ----
pred <- pred %>%
  dplyr::filter(replicates %in% mtdt$replicates)

rm(mtdt)

#---------------- PRELIMINAR PREDICTORS SELECTION / CLEANING ----------------

# Drop geometry ----
sel <- pred %>%
  st_drop_geometry() 


# Remove range ----
# Remove predictors containing "range"
range_cols <- grepl("_range", colnames(sel))

colnames(sel[which(range_cols)])


sel <- sel %>%
  dplyr::select(which(!range_cols))

rm(range_cols)
# Remove non weighted mean distances ----
sel <- sel %>%
  dplyr::select(-c(shore_dist_m_mean_log, mpa_dist_m_mean_log, port_dist_m_mean))
#--------------- PREDICTORS SELECTION v.1.0 ---------------- 

# Remove min and max cols ----
min_cols <- grepl("_min", colnames(sel))
max_cols <- grepl("_max", colnames(sel))
colnames(sel[which(min_cols)])
colnames(sel[which(max_cols)])

drop_cols <- min_cols | max_cols

sel <- sel[, !drop_cols]

rm(min_cols, max_cols, drop_cols)

# Remove day week and year cols ----
# remove cols containing "7j", "24h", "1y", "_year", "5years", "week", "day"
day_cols <- grepl("_7j", colnames(sel)) | grepl("_24h", colnames(sel)) | grepl("_1y", colnames(sel)) | grepl("_year", colnames(sel)) | grepl("5years", colnames(sel)) | grepl("week", colnames(sel)) | grepl("day", colnames(sel))

colnames(sel[which(day_cols)])

sel <- sel[, !day_cols]

rm(day_cols)

# Remove canyon associated cols (except distance) ----

canyon_cols <- grepl("canyon", colnames(sel)) & !grepl("canyon_dist_m_weight_log", colnames(sel))
colnames(sel[which(canyon_cols)])

sel <- sel[, !canyon_cols]


rm(canyon_cols)
# Remove not grouped habitat ----
hab_cols <- c("nb_habitat_per_km2", "main_habitat")

sel <- sel %>%
  dplyr::select(-all_of(hab_cols))

rm(hab_cols)

# Remove single habitat surface -----
# cols containing "clr"
clr_cols <- grepl("_clr", colnames(sel))

colnames(sel[which(clr_cols)])

sel <- sel %>%
  dplyr::select(-which(clr_cols))

rm(clr_cols)
# Remove sampling effort cols ----
samp_cols <- c("dist_seabed_depthsampling_log", "area_km2")

sel <- sel %>%
  dplyr::select(-all_of(samp_cols))

rm(samp_cols)

# Remove mpa_protection and mpa_fully ----
sel <- sel %>%
  dplyr::select(-mpa_protection) %>%
  dplyr::select(-mpa_fully)
# # Correlation matrix -----
# # Compute Spearman correlation matrix
# # Select numerical cols
# df <- sel %>%
#   dplyr::select(where(is.numeric))
# 
# # Compute correlation matrix
# cor_mat <- cor(df, method = "spearman", use = "pairwise.complete.obs")
# 
# # Open PNG device *before* plotting
# png("./figures/Predictors/Selection/corr_matrix_sel_v1.0_pre-selection_pred_tr_v1.3.png",
#     width = 2000, height = 2000, res = 300)
# 
# # Plot correlation matrix
# corrplot(cor_mat,
#          method = "color",
#          type = "upper",
#          tl.col = "black",
#          addCoef.col = "black",
#          tl.cex = 0.5,
#          cl.cex = 0.5,
#          number.cex = 0.5)
# 
# # Close device (this actually writes the file)
# dev.off()
# 
# # Results
# # Very high correlation > 0.9 between :
# # cop_analysed_sst_month_mean and temp_mean_1m --> we remove cop_analysed_sst_month_mean
# # roughness & tri & slope
# 
# # High correlation > 0.7 between :
# # canyon_dist and roughness, slope and tri
# 

# VIF ----
# Explanation : VIF can be used to detect collinearity (Strong correlation between two or more predictor variables). Collinearity causes instability in parameter estimation in regression-type models. The VIF is based on the square of the multiple correlation coefficient resulting from regressing a predictor variable against all other predictor variables. If a variable has a strong linear relationship with at least one other variables, the correlation coefficient would be close to 1, and VIF for that variable would be large. A VIF greater than 10 is a signal that the model has a collinearity problem. (source = usdm documentation)

# vifcor : first finds a pair of variables which has the maximum linear correlation (greater than the threshold; th), and exclude the one with a greater VIF. The procedure is repeated untill no pair of variables with a high corrrelation coefficient (grater than the threshold) remains.

# vifstep calculates VIF for all variables, excludes the one with the highest VIF (if it is greater than the threshold), repeat the procedure untill no variables with a VIF greater than th remains.


# Select numerical cols
df <- sel %>%
  dplyr::select(where(is.numeric))


# Print all VIF values
vif <- usdm::vif(df)
vif[order(-vif$VIF), ]

# Perform stepwise VIF selection
p_vifstep <- usdm::vifstep(df, th = 10, keep = NULL)

# Perform VIF correlation selection
p_vifcor <- usdm::vifcor(df, th = 0.7, keep = NULL, method = 'spearman') # remaining predictors with corr < th # default th is 0.9 # # make again with th = 0.7 method = spearman because we have non-normal data.

# Filter vif selected predictors
df_vifcor <- df[, p_vifcor@results$Variable]
df_vifstep <- df[, p_vifstep@results$Variable]

# Check removed predictors
setdiff(names(df), names(df_vifcor))
# [1] "temp_mean_1m"       "roughness_mean_log" "slope_mean_log"     "tri_mean_log"
# With Boats_density (pred_tr_v1.3) : 
# [1] "temp_mean_1m"       "gravity_mean_log"   "roughness_mean_log" "slope_mean_log"     "tri_mean_log"      

setdiff(names(df), names(df_vifstep))
# [1] "slope_mean_log" "tri_mean_log"
# With Boats_density (pred_tr_v1.3) : 
# [1] "temp_mean_1m"   "slope_mean_log" "tri_mean_log"  

rm(cor_mat, vif, p_vifcor, p_vifstep, df, df_vifcor, df_vifstep)

# Selection based on VIF and correlation results ----
# Remove roughness slope and tri -----
# slope and tri avec removed with vistep
# roughness is correlated to canyon distance and is removed by vifcor

sel <- sel %>%
  dplyr::select(-c(roughness_mean_log, slope_mean_log, tri_mean_log))


# Remove sst ----
# sst is highly correlated to temp_mean_1m and temp_mean is removed by vifcor. We prefer to keep mars3d temp that takes into account depth. 

sel <- sel %>%
  dplyr::select(-cop_analysed_sst_month_mean)



# # Correlation matrix after selection -----
# 
# # Compute Spearman correlation matrix
# df <- sel %>%
#   dplyr::select(where(is.numeric))
# 
# cor_mat <- cor(df, method = "spearman", use = "pairwise.complete.obs")
# 
# # Open PNG device *before* plotting
# png("./figures/Predictors/Selection/corr_matrix_sel_v1.0_post-selection_pred_tr_v1.3.png",
#     width = 2000, height = 2000, res = 300)
# 
# # Plot correlation matrix
# corrplot(cor_mat,
#          method = "color",
#          type = "upper",
#          tl.col = "black",
#          addCoef.col = "black",
#          tl.cex = 0.5,
#          cl.cex = 0.5,
#          number.cex = 0.5)
# 
# # Close device (this actually writes the file)
# dev.off()
# 
# # Clean
# rm(cor_mat, df)
# 
# 

# Add back geometry from pred ----
sel <- sel %>%
  left_join(pred %>% dplyr::select(replicates, geom),
            by = "replicates") %>%
  st_as_sf()



#--------------- EXPORTS ----------------
# # Export predictors_sel_v1.0.gpkg ----
# # based on mtdt_7_sel_v1.0.gpkg (755 obs) and predictors_tr_v1.1.gpkg (142 var + replicates)
# # selected predictors (16 + + x, y, replicates, geom): 
# colnames(sel)
# # [1] "grouped_main_habitat"       "x"                          "y"                          "replicates"                 "port_dist_m_weight"        
# # [6] "grouped_nb_habitat_per_km2" "bathy_mean"                 "wind_mean_1m"               "vel_mean_1m"                "temp_mean_1m"              
# # [11] "sal_mean_1m"                "northness"                  "canyon_dist_m_weight_log"   "mpa_dist_m_weight_log"      "shore_dist_m_weight_log"   
# # [16] "gravity_mean_log"           "tpi_mean_log"               "cop_chl_month_mean_log"     "eastness_log"               "geom"  
# 
# 
# 
# st_write(sel, "./data/processed_data/predictors/predictors_sel_v1.0.gpkg", delete_dsn = TRUE)
# 
# 
# # Export predictors_sel_v1.2.gpkg ----
# # based on mtdt_7_sel_v1.0.gpkg (755 obs) and predictors_tr_v1.2.gpkg (142 var + replicates --> x and y not transformed + negative log)
# # selected predictors (16 + + x, y, replicates, geom): 
# colnames(sel)
# # [1] "grouped_main_habitat"       "x"                          "y"                          "replicates"                 "port_dist_m_weight"        
# # [6] "grouped_nb_habitat_per_km2" "bathy_mean"                 "wind_mean_1m"               "vel_mean_1m"                "temp_mean_1m"              
# # [11] "sal_mean_1m"                "northness"                  "canyon_dist_m_weight_log"   "mpa_dist_m_weight_log"      "shore_dist_m_weight_log"   
# # [16] "gravity_mean_log"           "tpi_mean_log"               "cop_chl_month_mean_log"     "eastness_log"               "geom"    
# 
# 
# st_write(sel, "./data/processed_data/predictors/predictors_sel_v1.1.gpkg", delete_dsn = TRUE)
# 

# based on predictors_raw_v2.2
# transformations : Remove 0 variance predictors, Replace negative distance values by 0, CLT of habitat composition, Log when max value >/= to 10*median, Standard scale all --> same as predictors_tr_v1.1 BUT x and y are not scales nor log --> raw. + takes into account for negative values in log transformation
st_write(pred_tr, "./data/processed_data/predictors/predictors_tr_v1.2.gpkg", append = FALSE)


# # Export predictors_raw_sel_v1.3.gpkg ----
# # based on predictors_raw_v2.2 and predictors_sel_v1.3 
# # selects the same predictors and rows but from RAW predictors (for data explo)
# 
# pred_raw <- st_read("./data/processed_data/predictors/predictors_raw_v2.2.gpkg")
# predictors_sel_v1.3 <- st_read("./data/processed_data/predictors/predictors_sel_v1.3.gpkg")
# 
# sel_raw <- pred_raw %>%
#   dplyr::filter(replicates %in% predictors_sel_v1.3$replicates) %>% 
#   dplyr::select(c("replicates",
#                   "grouped_main_habitat",
#                   "x",
#                   "y",
#                   "aspect_mean",
#                   "tpi_mean",
#                   "port_dist_m_weight",
#                   "grouped_nb_habitat_per_km2",
#                   "bathy_mean",
#                   "wind_mean_1m",
#                   "vel_mean_1m",
#                   "temp_mean_1m",
#                   "sal_mean_1m",
#                   "canyon_dist_m_weight",
#                   "mpa_dist_m_weight",
#                   "shore_dist_m_weight",
#                   "gravity_mean",
#                   "cop_chl_month_mean",
#                   "geom"))
# 
# st_write(sel_raw, "./data/processed_data/predictors/predictors_raw_sel_v1.3.gpkg", delete_dsn = TRUE)

# Export predictors_sel_v1.3.gpkg ----
# Export predictors_sel_v1.4.gpkg ----
# based on predictors_tr_v1.3 (from predictors_raw_3.0 - with boats and habitat 2) 
# selected predictors : 
st_write(sel, "./data/processed_data/predictors/predictors_sel_v1.4.gpkg", delete_dsn = TRUE)









# Export predictors_sel_v.1.5.gpkg ----
# based on predictors_tr_v1.3 (from predictors_raw_3.0 - with boats and habitat 2)
# selected predictors : all
st_write(pred, "./data/processed_data/predictors/predictors_sel_v1.5.gpkg", delete_dsn = TRUE)
# Export predictors_sel_v.1.5_month.gpkg ----
# All predictors but only monthly temporal variables
sel <- pred
day_cols <- grepl("_7j", colnames(sel)) | grepl("_24h", colnames(sel)) | grepl("_1y", colnames(sel)) | grepl("_year", colnames(sel)) | grepl("5years", colnames(sel)) | grepl("week", colnames(sel)) | grepl("day", colnames(sel))

colnames(sel[which(day_cols)])

sel <- sel[, !day_cols]
names(sel)
rm(day_cols)

st_write(sel, "./data/processed_data/predictors/predictors_sel_v1.5_month.gpkg", delete_dsn = TRUE)

#----------------- AUTRES -----------------
#---- Visualise selected pred ----
method = 'vifcor' # 'vifcor'

if (method == 'vifstep') {
  pred_sel <- df_vifstep
  color <- "hotpink"
} else if (method == 'vifcor') {
  pred_sel <- df_vifcor
  color <- "skyblue"
}


# Correlation matrix
cm <- generate_correlation_matrix(data = df, exclude_column = NULL) # Spearman correlation

# Distance matrix
dm <- as.dist(1 - cm) # Dissimilarity measure (pearson = 1 -> 0 = identical, pearson = -1 -> -2 = maw dist)

# Dimension reduction for mapping/ visualise the data
mds_df <- as.data.frame(cmdscale(dm, k = 2))  # 2D scaling = PCA
mds_df <- cbind(Variable = rownames(mds_df), mds_df)  # Add Variable column
colnames(mds_df) <- c("Variable", "Dim1", "Dim2")  # Rename columns
rownames(mds_df) <- NULL  # Remove row names

# Add a column indicating whether the variable is in pred_sel
mds_df$Selected <- ifelse(mds_df$Variable %in% colnames(pred_sel), "Selected", "Not Selected")

# Create the plot
ggplot(mds_df, aes(x = Dim1, y = Dim2, label = Variable, color = Selected)) +
  geom_point(size = 2.5) +
  geom_text_repel(size = 2.5, max.overlaps = Inf) +  # Ensure all labels appear
  scale_color_manual(values = c("Selected" = color, "Not Selected" = "grey30")) +  # Highlight selected variables
  theme_minimal(base_size = 14) +  # Increase base font size
  theme(
    plot.title = element_text(size = 18, face = "bold", hjust = 0.5),
    legend.title = element_text(size = 14),
    legend.text = element_text(size = 12),
    panel.background = element_rect(fill = "white", color = "white"),  # White panel background
    plot.background = element_rect(fill = "white", color = "white")   # White plot background
  ) +
  labs(
    title = paste0("MDS with Predictors selected ", method),
    x = "Dimension 1",
    y = "Dimension 2",
    color = paste0("Variable Selection\n(", sum(mds_df$Selected == "Selected"), " Pred selected)")
  )






#---- Multivariate analyses ----
library(vegan)


# --- Data prep ---
occ <- read.csv("./data/processed_data/eDNA/occ_pooled_v1.1.csv")
str(occ)

# Remove rows not in sel
occ <- occ %>%
  dplyr::filter(replicates %in% sel$replicates)
occ <- occ[match(sel$replicates, occ$replicates), ]

# Build species matrices
occ_mat <- occ[, -1]
occ_pa  <- (occ_mat > 0) * 1

occ_sel <- occ_pa


# --- RDA ---
# RDA / dbRDA with Jaccard distance on species data
RDA2 <- vegan::capscale(
  occ_sel ~ northness +
    eastness +
    tpi_mean_log +
    port_dist_m_weight +
    bathy_mean +
    wind_mean_1m +
    vel_mean_1m +
    sal_mean_1m +
    cop_analysed_sst_month_mean +
    grouped_nb_habitat_per_km2 +
    canyon_dist_m_weight_log +
    mpa_dist_m_weight_log +
    shore_dist_m_weight_log +
    cop_chl_month_mean_log +
    Boat_density_month_log,
  data      = pred_sel,   # environmental variables live here
  dist      = "jaccard",
  na.action = na.omit,
  add       = TRUE
)

summary(RDA2)
plot(RDA2)


# --- AFC  ---
AFC <- FactoMineR::CA(occ_sel, graph = FALSE, axes = c(1, 2))
factoextra::fviz_ca_biplot(AFC,
                         axes = c(1, 2),
                         col.row = "blue",
                         col.col = "black",
                         repel = FALSE,
                         ellipse.type = "confidence",
                         addELLipse = TRUE)



# --- ACP  ---
PCA <- FactoMineR::PCA(pred_sel %>% dplyr::select(where(is.numeric)),
                        scale.unit = TRUE,
                        ncp = 5,
                        graph = FALSE)
# Get eigenvalues
eig.val <- factoextra::get_eigenvalue(PCA)
eig.val
# Draw screeplot
factoextra::fviz_eig(PCA, addlabels = TRUE, ylim = c(0, 50))
# Variable contributions to axes
var <- factoextra::get_pca_var(PCA)
# Corrplot of the cos2 of variables
corrplot::corrplot(var$cos2, is.corr = FALSE)
# Total cos2 of variables on Dim.1 to Dim.3
factoextra::fviz_cos2(PCA, choice = "var", axes = 1:3)
# Contributions of variables to PC1 :
factoextra::fviz_contrib(PCA, choice = "var", axes = 1, top = 10)
# Contributions of variables to PC2 :
factoextra::fviz_contrib(PCA, choice = "var", axes = 2, top = 10)
# Variable contributions to axes
factoextra::fviz_pca_var(
  PCA,
  col.var = "contrib",
  gradient.cols = c("#00AFBB", "#E7B800", "#FC4E07")
)
